+++ codex/views/browser/filters/field.py
@@ -5,6 +5,7 @@ from types import MappingProxyType
 from django.db.models.query_utils import Q
 
 from codex.views.browser.filters.group import GroupFilterView
+from codex.views.const import BROWSER_FILTER_KEYS
 from codex.views.settings.const import (
     CREDIT_PERSON_UI_FIELD,
     IDENTIFIER_TYPE_UI_FIELD,
@@ -18,32 +19,7 @@ _FILTER_REL_MAP = MappingProxyType(
         IDENTIFIER_TYPE_UI_FIELD: "identifiers__source",
     }
 )
-_FILTER_ATTRIBUTES: frozenset[str] = frozenset(
-    {
-        "age_rating",
-        "characters",
-        "country",
-        CREDIT_PERSON_UI_FIELD,
-        "critical_rating",
-        "decade",
-        "file_type",
-        "genres",
-        IDENTIFIER_TYPE_UI_FIELD,
-        "language",
-        "locations",
-        "monochrome",
-        "original_format",
-        "reading_direction",
-        "series_groups",
-        "stories",
-        STORY_ARC_UI_FIELD,
-        "tagger",
-        "tags",
-        "teams",
-        "universes",
-        "year",
-    }
-)
+_FILTER_ATTRIBUTES: frozenset[str] = frozenset(BROWSER_FILTER_KEYS)
 
 
 class ComicFieldFilterView(GroupFilterView):